event-ring: implement simplified versions of enqueue and full check
The Linux XHCI driver provides us with a single segment, which can fit
256 TRBs. We will have to wrap around to the start after writing 256
TRBs, flip the cycle state, and check the dequeue pointer to see whether
the ring is full. For now, we don't wrap around, so we can ignore the
cycle state toggle and dequeue pointer. Once we wrote 256 TRBs, the ring
will be unusuable.
The assumption is that we only need one or a few events to get the
driver to write to the command ring and ring door bell 0; this progress
takes priority over a full event ring implementation.