trb: add full implementation for the Link TRB
"full implementation" regarding the Command TRB. The Link TRB is the
only type of TRB that can appear on different ring kinds: on the command
ring and on transfer rings. The tranfer ring is not addressed in the
commit, but the LinkTrbData struct can be reused later for a TransferTrb
implementation.