xhci: handle Enable Slot Command
The Enable Slot Command expects us to return an available slot id for
the new device. Currently, the Linux driver sees two ports, which link
to the same port on our side. Thus, the driver tries to enable to
devices. The first attempt should get a valid slot id (1), the second
attempt we tell that there is no slot id available and cancel the
discovery process.
That there is only one available slot is part of the DeviceSlotManager
configuration, not particular to this handler implementation.