nusb: implement endpoint handling
The implementation is currently specific for USB storage (bulk endpoints
0x81 and 0x02). nusb uses the type system to differentiate bulk and
interrupt endpoints, as well as in and out endpoints. That is really
nice for explicitly handling endpoints in a userspace driver, but for us
the design complicates organizing enabled endpoints in a unified
collection.

We have to come up with a good idea here, maybe when we looked at
handling interrupt endpoints, too.