smoke-test: fix usbvfiod-socket race condition
The usbvfiod service is started before the cloud-hypervisor service,
however it is possible that usbvfiod has not yet created the socket when
cloud-hypervisor tries to access it.

The race condition has not yet happened in CI testing, but happens
sometimes when running locally and happens often when running
interactively.

This commit introduces a simple bandaid. When the cloud-hypervisor was
to quick, systemd will restart the service again two seconds later, at
which point the socket is pretty much guaranteed to be available
(otherwise just another restart happens).

Luckily, `machine.wait_for_unit("cloud-hypervisor.service")` does not
fail when the service fails temporarily and is restarted.