nusb,xhci: handle transfer IN through worker
There are changes to two functionalities.

(1) enable_endpoint: Instead of opening and storing the endpoint with
the nusb interface handles, we launch the background worker thread and
store the mpsc::Sender that allows us to wake up the thread.

(2) transfer_in: Instead of the whole handling logic (retrieve TRB,
initiate transfer on real endpoint, emulate DMA), we simply wake up the
worker thread that runs all the handling in the background.