xhci: send port status change event on device hotplug
When we attach a device while the controller is running, just changing
the portsc is not enough to inform the driver about the new device. We
have to send a port status change to the driver.
However, if we attach device at startup, before the controller is
running, we must not send interrupts as the driver does not expect any
yet.