test.nix: add a test for a single interrupt endpoint via qemu device usb-kbd
UDEV Rule: In some cases the test behaved flaky due to the symlink missing. After starting usbvfiod the symlink (for some reason) was consumed. As far as I understand that usually happens when the originally linked device is unavailable or in some other way changed and the link is removed.
When the UDEV rule is triggered not only on add but also on change whatever happens to the device will trigger the rule and create a new symlink that enables usbvfiod to start.