test.nix: replace testing with a function and add debug arg
Previously we had duplicate code because of things like "usbvfiod will need a different ExecStart string" and some more boilerplate in every Test.
This commit adds a function to create tests for usbvfiod. Using a attrs as argument only "different things" are included in the tests definition. All four previous tests are possible with this new function.

This also removes the old tests and provides a drop in using the new function.

Additionally a debug argument was added to have the option to have production environment like logs (i.e. disable the usual debug logs of a dev environment).
The four tests kept their previous log verbosity.