hotplug: path resolution and bus/dev extraction
When we receive a symlink to a USB device file, we can run into
problematic behavior. The problem exists currently already and will also
exist with hot-attaching:
- currently: when a symlink is passed to --device, we open the file to
reset the USB device and then have to open the file again. When the
symlink changes in between, we reset a different device than what we
finally open.
- with hot-attach: the remote application opens a device file, resets,
opens the file again and sends the file descriptor over a socket to
usbvfiod. The same problem occurs when a symlink is used and the symlink
changes target in between.

Instead, we want to to canonicalize the symlink once (to what should be
a /dev/bus/usb/xxx/yyy path) and perform the reset and opening with the
device-file path directly.