remote: print command on attachment
When the caller of an attachment command uses a symlink to pass the
driver file, he might not be aware which bus/device number the device
has. The list command lists attached devices by their bus/device number,
and bus/device numbers need to be specified for device removal. The
newly printed message let's the caller know how to identify the device
from now on.