hotplug: fix reading incorrect response when socket is closed
When working on implementing the detach command, I noticed the following
weird situation:
- Remote side for detach command fully implemented
- usbvfiod side accepted the command, but ran into a todo!()
- thus, usbvfiod never sent back a response over the socket
Still, the remote tool printed that it received a SuccessfulOperation
response (0x00). The description of UnixStream::read describes what
happens: the read encounters EOF (because the todo!() crash closed the
socket), and, in this case, it does not return an error but Ok(0).
Our Response::receive_from_socket does not check the value and returns
the byte in the buffer, which is still 0x00 from initialization.

Changing to read_exact fixes the problem as we now receive and forward
an error when the single byte cannot be read successfully.