nusb: let endpoints mirror the spec graphics with 32 entries instead of 30
We never needed the index 0 and 1 and could save on array length.

This made things harder to understand for me. This is probably a
subjective design decision and there is not really a right or wrong: We
reference the xhci spec often enough that I think we should, wherever
possible, implement things as they are shown in graphics etc. In this
case this means a full 32 length array of endpoints. Even if we might
not use index 0 and 1 currently, any developer starting out will likely
have the spec open to better understand the code.