test.nix: add timeout to all x.succeed() commands
The x.succeed() method does not have the 900 second timeout the
x.execute() or x.wait_until_succeeds() methods do. Because of that we
ran into global timeouts that do not trigger the print of the guest
machine logs. With this change the individual x.succeed() should trigger
a timeout and then print the guest logs before failing the test.

To keep the custom python class as close to the original test framework
as possible I decided against setting a default timeout and instead set
one for each call.