xhci: address ports by their id
so far, we stored devices and portsc register in arrays and addressed
them with their "index" (starting at 0). We recently changed the
endpoint addressing to using the id as the index (and having some empty
space at the start of the array) to avoid confusion and off-by-X errors.
This commit makes the same change for ports.